+++ Contents/mods/Evolving Traits World/42.19/media/lua/shared/ETW_MarkDynamicTraits.lua
@@ -34,6 +34,7 @@ local function registerDynamicTraits()
 		ETW_Registry.traits.GUN_ENTHUSIAST,
 		ETW_Registry.traits.TERMINATOR,
 		ETW_Registry.traits.GYM_RAT,
+		ETW_Registry.traits.COUCH_POTATO,
 		ETW_Registry.traits.HOARDER,
 		ETW_Registry.traits.HOME_COOK,
 		ETW_Registry.traits.HOMICHLOPHILE,