+++ Contents/mods/Evolving Traits World/42.19/media/lua/shared/Types/ETW_ModDataType.lua
@@ -76,6 +76,7 @@
 --- @field SmokeSystem SmokeSystem
 --- @field TransferSystem TransferSystem
 --- @field BloodlustSystem BloodlustSystem
+--- @field GymTraitsSystemCounter number
 --- @field AnimalsSystem AnimalsSystem
 
 --- @class TVJunkieSystem